Message_And_Video_Service/Message_Video/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.2
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
av==13.1.0
bidict==0.23.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
environ==1.0
eventlet==0.39.0
exceptiongroup==1.2.2
frozenlist==1.5.0
google-crc32c==1.6.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
ifaddr==0.2.0
incremental==24.7.2
mongoengine==0.29.1
msgpack==1.1.0
multidict==6.1.0
propcache==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyee==12.1.1
PyJWT==2.10.1
pylibsrtp==0.11.0
pymongo==4.11.1
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
